+++ docs/mise_a_jour_donnees_traces_fcu.md
@@ -902,9 +902,9 @@ yarn cli download-network futurNetwork
 
 # voir génération custom via geojson pour les réseaux de chaleur via doc !
 yarn cli tiles:generate-from-file reseaux_de_chaleur.geojson reseaux_de_chaleur_tiles 0 14
-yarn cli fill-tiles coldNetwork 0 14
-yarn cli fill-tiles zoneDP 0 14
-yarn cli fill-tiles futurNetwork 0 14
+yarn cli tiles:fill coldNetwork 0 14
+yarn cli tiles:fill zoneDP 0 14
+yarn cli tiles:fill futurNetwork 0 14
 
 # copie en dev
 ./scripts/copyLocalTableToRemote.sh dev reseaux_de_chaleur --data-only
