+++ scripts/fillTiles.ts
@@ -1,6 +1,6 @@
 import { type DatabaseSourceId } from '@/server/services/tiles.config';
 
-import { fillTiles } from './utils/tiles';
+import { fillTiles } from './tiles/utils';
 
 if (process.argv.length !== 5 && process.argv.length !== 6) {
   console.info('Usage: npx tsx scripts/fillTiles.ts table zoomMin zoomMax [index]');
